+++ src/main/java/com/example/SynClock/services/JwtUserDetailsService.java
@@ -1,9 +1,9 @@
 package com.example.SynClock.services;
 
 import com.example.SynClock.model.User;
+import com.example.SynClock.repositories.UserRepository;
 import com.example.SynClock.security.JwtUserDetails;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -19,7 +19,7 @@ public class JwtUserDetailsService {
 //        User user = userRepository.findByUsername(username);
 //        return JwtUserDetails.create(user);
 //    }
-    public UserDetails loadUserByid(Integer uuid){
+    public UserDetails loadUserByid(Long uuid){
         User user = userRepository.findById(uuid).get();
         return JwtUserDetails.create(user);
     }
